3.1-drf-intro/smart_home/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
Django==5.1.6
djangorestframework==3.15.2
flake8==7.1.2
idna==3.10
iniconfig==2.0.0
mccabe==0.7.0
packaging==24.2
pillow==11.1.0
pluggy==1.5.0
psycopg2-binary==2.9.10
pycodestyle==2.12.1
pyflakes==3.2.0
pytest==8.3.4
pytest-django==4.10.0
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.3
tzdata==2025.1
urllib3==2.3.0
